Add a combo box to select recording format between raw and SigMF
I am not super happy with the current implementation. It seems like it
might be cleaner to factor out a recording class hierarchy where raw
and sigmf would implement a common interface. This could then also be
used to record more metadata during the recording (e.g. when we tune the
radio to a different frequency, we should start a new capture in sigmf).
However, that might be a lot of boilerplate since these are the only two
formats that we're looking to support for now.